etl/requirements.txt
pandas==2.0.3
sqlalchemy==2.0.21
pymysql==1.1.0
python-dotenv==1.0.0
openpyxl==3.1.2